# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "UMS API DOCUMENTATION",
        "description": "A simple laravel user management system for users",
        "contact": {
            "email": "admin@admin.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://my-default-host.com",
            "description": "Demo API Server"
        }
    ],
    "paths": {
...
        "api/v1/register": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Register user",
                "description": "This can only be done by an guest.",
                "operationId": "createUser",
                "requestBody": {
                    "description": "Create user object",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateUser"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/User"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Unprocessible Entity"
                    }
                }
            }
        },
        "login": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "description": "Login in a User",
                "operationId": "userLogin",
                "requestBody": {
                    "description": "Updated user object",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UserLogin"
                            }
                        }
                    }
                },
                "responses": {
                    "401": {
                        "description": "Unathorize Access"
                    },
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/User"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/{email}/verify/{verification_code}": {
            "get": {
                "tags": [
                    "Auth"
                ],
                "summary": "Verify a user account.",
                "description": "Verify a User",
                "operationId": "verifyUser",
                "parameters": [
                    {
                        "name": "email",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "verification_code",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "401": {
                        "description": "Unathorize Access"
                    },
                    "200": {
                        "description": "successful operation"
                    }
                }
            }
        },
        "/{email}/resend_verification_code": {
            "get": {
                "tags": [
                    "Auth"
                ],
                "summary": "Resend verification link.",
                "description": "Resend verification link",
                "operationId": "resendVerification",
                "parameters": [
                    {
                        "name": "email",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "401": {
                        "description": "Unathorize Access"
                    },
                    "200": {
                        "description": "successful operation"
                    }
                }
            }
        },
        "/{email}/forgot_password": {
            "get": {
                "tags": [
                    "Auth"
                ],
                "summary": "Forgot Password.",
                "description": "Sends mail containing token to users who forgot their password",
                "operationId": "forgotPassword",
                "parameters": [
                    {
                        "name": "email",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "401": {
                        "description": "Unathorize Access"
                    },
                    "200": {
                        "description": "successful operation"
                    }
                }
            }
        },
        "/{email}/reset_password": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Reset User password.",
                "description": "Reset a user's password",
                "operationId": "resetUserPassword",
                "parameters": [
                    {
                        "name": "passwordreset",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "$ref": "#/components/schemas/PasswordReset"
                        }
                    }
                ],
                "responses": {
                    "401": {
                        "description": "Unathorize Access"
                    },
                    "200": {
                        "description": "successful operation"
                    }
                }
            }
        },
        "/logout": {
            "get": {
                "tags": [
                    "Auth"
                ],
                "description": "Logout a user",
                "operationId": "userLogout",
                "responses": {
                    "401": {
                        "description": "Unathorize Access"
                    },
                    "200": {
                        "description": "successful operation"
                    }
                },
                "security": [
                    {
                        "APP_auth": [
                            "read:user"
                        ]
                    }
                ]
            }
        },
...
        "/users/{uuid}": {
            "patch": {
                "tags": [
                    "User"
                ],
                "summary": "Updated user",
                "description": "This can only be done by an auth users.",
                "operationId": "updateUser",
                "parameters": [
                    {
                        "name": "uuid",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Updated user object",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateUser"
                            }
                        }
                    }
                },
                "responses": {
                    "400": {
                        "description": "bad request"
                    },
                    "404": {
                        "description": "User not found"
                    },
                    "422": {
                        "description": "Unprocessible Entity"
                    },
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/User"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
...
            "User": {
                "properties": {
                    "firstname": {
                        "type": "string"
                    },
                    "lastname": {
                        "type": "string"
                    },
                    "middlename": {
                        "type": "string"
                    },
                    "email": {
                        "type": "string"
                    },
                    "phone_number": {
                        "type": "integer"
                    },
                    "profile_url": {
                        "type": "string"
                    },
                    "password": {
                        "type": "string"
                    },
                    "created_at": {
                        "type": "string"
                    },
                    "deleted_at": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "CreateUser": {
                "properties": {
                    "firstname": {
                        "type": "string"
                    },
                    "lastname": {
                        "type": "string"
                    },
                    "middlename": {
                        "type": "string"
                    },
                    "email": {
                        "type": "string"
                    },
                    "phone_number": {
                        "type": "integer"
                    },
                    "password": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "UpdateUser": {
                "required": [
                    "uuid"
                ],
                "properties": {
                    "firstname": {
                        "type": "string"
                    },
                    "lastname": {
                        "type": "string"
                    },
                    "email": {
                        "type": "string"
                    },
                    "phone": {
                        "type": "integer"
                    },
                    "password": {
                        "type": "string"
                    },
                    "created_at": {
                        "type": "string"
                    },
                    "deleted_at": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "UserLogin": {
                "required": [
                    "email",
                    "password",
                    "role_id"
                ],
                "properties": {
                    "email": {
                        "type": "string"
                    },
                    "password": {
                        "type": "string"
                    },
                    "role_id": {
                        "type": "integer"
                    }
                },
                "type": "object"
            },
            "PasswordReset": {
                "required": [
                    "password",
                    "token"
                ],
                "properties": {
                    "password": {
                        "type": "string"
                    },
                    "token": {
                        "type": "string"
                    }
                },
                "type": "object"
            }
        }
    }
}